refactor : 기존 ncp 활용 로직 경로이동
